Replace upload route with state
