+++ www/src/redux/substores/student/slices/resumeReviewSlice.ts
@@ -1,4 +1,4 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import { ResumeReview } from '../../../../util/serverResponses';
 import getMyResumeReviews from '../thunks/getMyResumeReviews';
@@ -6,17 +6,23 @@ import getMyResumeReviews from '../thunks/getMyResumeReviews';
 type InitialState = {
     resumeReviews: ResumeReview[];
     isLoading: boolean;
+    isUploading: boolean;
 };
 
 const initialState: InitialState = {
     resumeReviews: [],
     isLoading: false,
+    isUploading: false,
 };
 
 export const resumeReviewSlice = createSlice({
     name: 'resumeReview',
     initialState,
-    reducers: {},
+    reducers: {
+        setIsUploadingResume: (state, action: PayloadAction<boolean>) => {
+            state.isUploading = action.payload;
+        },
+    },
     extraReducers: (builder) => {
         builder.addCase(getMyResumeReviews.pending, (state) => {
             state.isLoading = true;
@@ -26,5 +32,7 @@ export const resumeReviewSlice = createSlice({
         });
     },
 });
+
+export const { setIsUploadingResume } = resumeReviewSlice.actions;
 
 export default resumeReviewSlice.reducer;
