+++ www/src/util/constants.ts
@@ -8,7 +8,6 @@ export const COMPE_CLUB = 'CompE Club';
 
 // Routes
 export const RESUME_REVIEW_ROUTE = '/resume-review';
-export const UPLOAD_RESUME_REVIEW_ROUTE = '/resume-review/upload';
 export const MOCK_INTERVIEW_ROUTE = '/mock-interview';
 export const COMMUNITY_ROUTE = '/community';
 export const REGISTER_ROUTE = '/register';
